+++ backend/business-partner-agent/src/main/resources/org/hyperledger/bpa/i18n/messages.properties
@@ -35,7 +35,6 @@ api.issuer.credential.send.offer.wrong.state=Wrong exchange state expected: prop
 api.issuer.credential.send.offer.wrong.creddef=Provided credential definition id is not configured
 api.issuer.credential.send.offer.wrong.schema=Provided schema id is not configured
 api.issuer.credential.send.not.supported=This operation is not supported for json-ld credentials
-api.schema.credential.document.conversion.failure=Only documents that are based on a schema can be converted into a credential
 api.issuer.creddef.already.exists=Credential definition for schema: {id} already exists with tag: {tag}
 api.issuer.creddef.ledger.failure=Credential Definition not created; could not complete request with ledger
 api.issuer.creddef.in.use=Credential Definition cannot be deleted, it has been used to issue credentials
