+++ backend/pom.xml
@@ -54,7 +54,7 @@
         <skip.docker.build>true</skip.docker.build>
         <!-- Dependency Versions -->
         <jackson.version>2.13.3</jackson.version>
-        <log4j2.version>2.17.2</log4j2.version>
+        <log4j2.version>2.18.0</log4j2.version>
         <lombok.version>1.18.24</lombok.version>
         <lombok.maven.plugin.version>1.18.20.0</lombok.maven.plugin.version>
         <micronaut.version>3.5.1</micronaut.version>
@@ -66,7 +66,7 @@
         <org.mapstruct.version>1.4.2.Final</org.mapstruct.version>
         <pmd.version>6.47.0</pmd.version>
         <spotbugs.version>4.7.1</spotbugs.version>
-        <testcontainers.version>1.17.2</testcontainers.version>
+        <testcontainers.version>1.17.3</testcontainers.version>
         <!-- Plugin Versions -->
         <license-maven-plugin.version>4.1</license-maven-plugin.version>
         <maven.compiler.plugin.version>3.10.1</maven.compiler.plugin.version>
